+++ packages/thirdweb/src/utils/encoding/helpers/concat-hex.ts
@@ -5,6 +5,7 @@ import type { Hex } from "./is-hex.js";
  *
  * @param values - An array of hexadecimal values to concatenate.
  * @returns The concatenated hexadecimal value.
+ * @utils
  */
 export function concatHex(values: readonly Hex[]): Hex {
   return `0x${(values as Hex[]).reduce(