+++ legacy_packages/cli/CHANGELOG.md
@@ -1,5 +1,14 @@
 # thirdweb
 
+## 0.14.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @thirdweb-dev/sdk@4.0.90
+  - @thirdweb-dev/wallets@2.5.30
+  - @thirdweb-dev/auth@4.1.88
+
 ## 0.14.2
 
 ### Patch Changes